+++ racker/cli.py
@@ -10,8 +10,9 @@ import click
 
 from postroj.api import pull_curated_image
 from postroj.container import PostrojContainer
-from postroj.exceptions import ProvisioningError
+from postroj.exceptions import ProvisioningError, InvalidImageReference
 from postroj.util import boot, subprocess_get_error_message
+from postroj.winrunner import WinRunner
 from racker.image import ImageLibrary
 
 logger = logging.getLogger(__name__)
@@ -47,10 +48,11 @@ def racker_pull(ctx, name: str):
 @click.option("--interactive", "-i", is_flag=True)
 @click.option("--tty", "-t", is_flag=True)
 @click.option("--rm", is_flag=True)
+@click.option("--platform", type=str, required=False)
 @click.argument("image", type=str, required=False)
 @click.argument("command", nargs=-1, type=str)
 @click.pass_context
-def racker_run(ctx, interactive: bool, tty: bool, rm: bool, image: str, command: str):
+def racker_run(ctx, interactive: bool, tty: bool, rm: bool, platform: str, image: str, command: str):
     """
     Spawn a container and run a command on it.
     Aims to be compatible with `docker run`.
@@ -84,6 +86,45 @@ def racker_run(ctx, interactive: bool, tty: bool, rm: bool, image: str, command:
     if interactive or tty:
         use_pty = True
 
+    # Use a different subsystem for running Windows containers on Linux or
+    # macOS.
+    if platform is None:
+        pass
+
+    elif platform == "windows/amd64":
+        logger.info(f"Preparing runtime environment for platform {platform} and image {image}")
+        try:
+            runner = WinRunner(image=image)
+        except InvalidImageReference as ex:
+            raise SystemExit(ex.returncode)
+        try:
+            with redirect_stdout(sys.stderr):
+                with redirect_stderr(sys.stderr):
+                    runner.setup()
+                    runner.start()
+        except subprocess.CalledProcessError as ex:
+            message = subprocess_get_error_message(exception=ex)
+            logger.critical(f"Launching container failed. {message}")
+            # subprocess_forward_stderr_stdout(exception=ex)
+            raise SystemExit(ex.returncode)
+
+        logger.info(f"Invoking command '{command}' on {image}")
+        try:
+            runner.run(command, interactive=interactive, tty=tty)
+        except subprocess.CalledProcessError as ex:
+            message = subprocess_get_error_message(exception=ex)
+            logger.critical(f"Running command in Windows container on image {image} failed. {message}")
+            raise SystemExit(ex.returncode)
+        return
+
+    else:
+        raise NotImplementedError(f'Runtime for platform "{platform}" not supported yet')
+
+    if sys.platform != "linux":
+        raise NotImplementedError(f"Unable to launch Linux systems on non-Linux machines yet, "
+                                  f"please use the Vagrant setup")
+
+    # Acquire filesystem image.
     # TODO: Add more advanced image registry, maybe using `docker-py`,
     #       resolving image names from postroj-internal images, Docker Hub,
     #       GHCR, etc.