+++ CHANGES.rst
@@ -8,7 +8,7 @@ in progress
 
 - Remove support for Debian 9 "stretch"
 - Added support for Debian 13 "trixie"
-
+- Added support for Ubuntu 24 "Noble Numbat"
 
 2023-02-14 0.3.0
 ================